==> Code/02_Simple_Regression.Rmd <==
---
title: "02_Simple_Regression"
author: "Abraham Ybarra"
date: '2023-02-03'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Load libraries

```{r}
library(tidyverse)
library(here)
library(afex)
```

## Read in data

```{r}
######## Read in F21 (Ashley) Data ############
studentSummary <- read.csv(paste0(here(),"/clean_data/F21_Ashley/studentSummary.csv"), header = TRUE)
late <- read.csv(paste0(here(),"/clean_data/F21_Ashley/lateAssignments.csv"), header = TRUE)
missing <- read.csv(paste0(here(),"/clean_data/F21_Ashley/missingAssignments.csv"), header = TRUE)
grades <- read.csv(paste0(here(),"/clean_data/F21_Ashley/gradebook.csv"), header = TRUE)
```

## Plot functions

```{r}
######## Create scatter plot (x vs y) function ############
scatter_plot <- function(title, xvar, yvar, ylim = c(0,100), model = NULL) {
  p <- ggplot(studentSummary) +
  
    # aesthetics
    geom_point(aes_string(x = xvar, y = yvar)) +
    geom_abline(intercept = coef(model)[1],
                slope = coef(model)[2]) +
    
    ylim(ylim) +
    
    # title and label axes
    ggtitle(paste(xvar,"vs",yvar)) +
    xlab(xvar) + 
    ylab(yvar) +
    theme(title = element_text(size = 15),
          text = element_text(size = 11),
          axis.text.x = element_text(margin = margin(5,0,20,0)),
          axis.text.y = element_text(margin = margin(0,5,0,20)),
          plot.title = element_text(margin = margin(0,0,20,0), hjust = 0.5))
  
  ggsave(paste0(here(),"/Plots/",title,".jpeg"))
}
```

## Plots

```{r}
######## Scatter plots of key variables vs. overall course grade ############
scatter_plot(title = "On time pct vs grade", xvar = "Assignment.on.time.percent", yvar = "Overall.course.grade", ylim = c(50,100))
scatter_plot(title = "Page view vs grade", xvar = "Page.Views", yvar = "Overall.course.grade", ylim = c(50,100))
scatter_plot(title = "Participations vs grade", xvar = "Participations", yvar = "Overall.course.grade", ylim = c(50,100))
```


## Analyses

```{r}
######## Mean center predictor variables ############
studentSummary <- studentSummary %>%
  mutate(across(c(Assignment.on.time.percent, Page.Views, Participations), ~ .x - mean(.x)))

set_sum_contrasts()

######## Create models ############
m.1 <- lm(Overall.course.grade ~ Assignment.on.time.percent,  data = studentSummary)
summary(m.1)

m.2 <- lm(Overall.course.grade ~ Page.Views,  data = studentSummary)
summary(m.2)

m.3 <- lm(Overall.course.grade ~ Participations,  data = studentSummary)
summary(m.3)

m.4 <- lm(Overall.course.grade ~ Assignment.on.time.percent * Page.Views,  data = studentSummary)
summary(m.4)

m.5 <- lm(Overall.course.grade ~ Assignment.on.time.percent * Page.Views * Participations,  data = studentSummary)
summary(m.5)
```


## Post-analysis plots

```{r}
######## Add model predictions to dataset ############
studentSummary <- studentSummary %>%
  mutate(m.1.pred = predict(m.1),
         m.2.pred = predict(m.2),
         m.3.pred = predict(m.3),
         m.4.pred = predict(m.4),
         m.5.pred = predict(m.5))

######## Scatter plots of key variables vs. overall course grade ############
scatter_plot(title = "REG On time pct vs grade", xvar = "Assignment.on.time.percent", yvar = "Overall.course.grade", ylim = c(50,100), 
             model = m.1)
scatter_plot(title = "REG Page view vs grade", xvar = "Page.Views", yvar = "Overall.course.grade", ylim = c(50,100),
             model = m.2)
scatter_plot(title = "REG Participations vs grade", xvar = "Participations", yvar = "Overall.course.grade", ylim = c(50,100),
             model = m.3)
```
