==> Code/04_Descriptive_Stats.Rmd <==
---
title: "04_Descriptive_Stats"
author: "Abraham Ybarra"
date: '2023-02-06'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# This file creates descriptive plots (e.g., histograms, box plots) for various procrastination indices and outcome variables of interest

## Load libraries

```{r}
library(tidyverse)
library(here)
```

## Read in data

```{r}
######## Read in F21 (Ashley) Data ############
studentSummary_21A <- read.csv(paste0(here(),"/final_data/F21_Ashley/studentSummary.csv"), header = TRUE)
assignments_21A <- read.csv(paste0(here(),"/final_data/F21_Ashley/assignments.csv"), header = TRUE, numerals = "no.loss")
```

## Histogram plotting function

```{r}
######## Histogram plotting function ############
plot_histogram <- function(df, xvar, binwidth, xlim = c(-10,10)) {
  p <- ggplot(df) +
    
    # aesthetics
    geom_histogram(aes_string(x = xvar), binwidth = binwidth, color = "blue", fill = "blue", alpha = .5) +
      
    # title and label axes
    ggtitle(paste("Distribution of", xvar)) +
    xlab(xvar) +
    
    xlim(xlim) +
    
    # Font size 
    theme_classic() +
    theme(plot.title = element_text(size = 18, margin = margin(0,0,20,0)),
          axis.title.x = element_text(size = 18, margin = margin(5,0,10,0)),
          axis.title.y = element_text(size = 18, margin = margin(0,5,0,10)),
          axis.text.x = element_text(size = 16, margin = margin(5,0,10,0)),
          axis.text.y = element_text(size = 16, margin = margin(0,5,0,10))) 
    
  ggsave(paste0(here(),"/Plots/Descriptive Stats/Distribution of ",xvar,".jpeg"))
}
```


## Plot histograms 
```{r}
######## Histograms for distribution of each variable ############
plot_histogram(df = studentSummary_21A, xvar = "OVERALL_COURSE_GRADE", binwidth = 5, xlim = c(0,100))
plot_histogram(df = studentSummary_21A, xvar = "FINAL_EXAM_SCORE", binwidth = 5, xlim = c(0,100))
plot_histogram(df = studentSummary_21A, xvar = "PCT_ON_TIME", binwidth = .05, xlim = c(0,1))
plot_histogram(df = studentSummary_21A, xvar = "PCT_LATE", binwidth = .05, xlim = c(0,.5))
plot_histogram(df = studentSummary_21A, xvar = "LOG_PCT_LATE", binwidth = 1, xlim = c(-7,2))
plot_histogram(df = studentSummary_21A, xvar = "MEAN_HOUR_DIFF", binwidth = 5, xlim = c(-100,100))
plot_histogram(df = studentSummary_21A, xvar = "MEAN_HOUR_DIFF_Z", binwidth = .25, xlim = c(-2,2))
plot_histogram(df = studentSummary_21A, xvar = "MEAN_PI", binwidth = .1, xlim = c(0,1))
plot_histogram(df = studentSummary_21A, xvar = "LOG_MEAN_PI", binwidth = .5, xlim = c(-5,1))
plot_histogram(df = assignments_21A, xvar = "HOUR_DIFF", binwidth = 5, xlim = c(-200,200))
plot_histogram(df = assignments_21A, xvar = "HOUR_DIFF_NORM", binwidth = .1, xlim = c(-1,1))
plot_histogram(df = assignments_21A, xvar = "HOUR_DIFF_Z", binwidth = .25, xlim = c(-3,3))
plot_histogram(df = assignments_21A, xvar = "LOG_HOUR_DIFF", binwidth = .25, xlim = c(2,6))
plot_histogram(df = assignments_21A, xvar = "YJ_NEG_2", binwidth = .25, xlim = c(-1.5,1.5))
plot_histogram(df = assignments_21A, xvar = "YJ_NEG_1", binwidth = .25, xlim = c(-1.5,1.5))
plot_histogram(df = assignments_21A, xvar = "YJ_0", binwidth = 1)
plot_histogram(df = assignments_21A, xvar = "YJ_POS_1", binwidth = 1)
plot_histogram(df = assignments_21A, xvar = "YJ_POS_2", binwidth = 1)
```

## Function: plot variables by assignment type

```{r}
######## Plot HOUR_DIFF by assignment type: DENSITY ############
plot_wrap <- function(df, xvar, binwidth = 5, xlim = c(-1,1)) {

  ggplot(df) +
      
    # aesthetics
    geom_histogram(aes(x = !!sym(xvar), y = binwidth*..density..), binwidth = binwidth, color = "blue", fill = "blue", alpha = .5) +
    facet_wrap(~ ASSIGNMENT_TYPE) +
    theme_classic() +
    
    xlim(xlim) +
        
    # title and label axes
    ggtitle(paste("Distribution of",xvar)) +
    xlab(xvar) +
    ylab("Density")
      
    # Font size
    theme(text = element_text(size = 12),
          axis.text.x = element_text(size = 12, margin = margin(5,0,20,0)),
          axis.text.y = element_text(size = 12, margin = margin(0,5,0,20)),
          plot.title = element_text(size = 16, margin = margin(0,0,20,0), hjust = 0.5)) +
    theme_classic()
      
  ggsave(paste0(here(),"/Plots/Descriptive Stats/", xvar, " by ASSIGNMENT_TYPE.jpeg"))
}
```

## Plot wraps

```{r}
plot_wrap(df = assignments_21A, xvar = "HOUR_DIFF", binwidth = 5, xlim = c(-100,100))
plot_wrap(df = assignments_21A, xvar = "HOUR_DIFF_NORM", binwidth = .1, xlim = c(-1,1))
plot_wrap(df = assignments_21A, xvar = "HOUR_DIFF_Z", binwidth = .25, xlim = c(-3,3))
```

## Boxplot function

```{r}
######## Plot HOUR_DIFF by assignment type: DENSITY ############
plot_box <- function(df, xvar, yvar) {

  ggplot(df, aes_string(x = xvar, y = yvar)) +
      
    # aesthetics
    geom_boxplot(alpha = .2, outlier.shape = NA) +
    stat_boxplot(coef = 5) + 
    theme_classic() +
        
    # title and label axes
    ggtitle(paste("Boxplot of",xvar)) +
    xlab(xvar) +
    ylab(yvar) +
      
    # Font size
    theme(text = element_text(size = 12),
          axis.text.x = element_text(size = 12, margin = margin(5,0,20,0)),
          axis.text.y = element_text(size = 12, margin = margin(0,5,0,20)),
          plot.title = element_text(size = 16, margin = margin(0,0,20,0), hjust = 0.5)) +
    theme_classic()
      
  ggsave(paste0(here(),"/Plots/Descriptive Stats/Boxplot of ", xvar, ".jpeg"))
}
```

## Plot boxplots

```{r}
assignments_21A <- assignments_21A %>%
  mutate(PI_FCT = as_factor(PI), .after = PI)

plot_box(df = assignments_21A, xvar = "PI_FCT", yvar = "SCORE_100")
```
